Consolidate IsRemoteURI in utils, extract ResolveRelativePaths, remove dead IsLocalPath
